oem+drivers: fix wrong order or lower layers when building overlays
When doing an offline install, ubuntu-drivers would sometimes list a
package that is available in the archive but not present in the pool.

This is not something we would expect since we run apt-get update (with
only the pool configured when offline) in the install tree.

However, it turned out that we create the overlay with the lower layers
specified in the wrong order - which essentially makes APT indexes
visible in the source tree also visible in the OEM/third-party driver
overlay.

When calling setup_overlay(lowers=[a, b, c]), Subiquity invokes mount
with lowerdir=c:b:a (in the reverse order).

This means that c is top, b is middle and a is bottom.

For the OEM and third-party drivers, we build overlays that are based
on:

 * the source tree
 * the configured tree
 * the install tree

Unfortunately, we were doing the opposite. Fixed by reversing the order of
the lower layers.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>