apport: show name of curtin stage when an error occurs
When an error occurs during the execution of a "$ curtin install"
invocation, the associated bug report would be titled:

  "install failed crashed with CalledProcessError"

This is not very user friendly and it does not immediately give any
insight on:

With the number of reports that we receive these days, I think
classifying these reports would make our life a bit easier, and possibly
make the users understand better what's going on.

This change adds a wat to detect if a bug report is the result of a
"$ curtin install" invocation failing. If it is, subiquity sets the
title of the bug report accordingly. Examples:

 * "partitioning crashed with CurtinInstallError"
 * "extract crashed with CurtinInstallError"
 * "curthooks crashed with CurtinInstallError"

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>