drivers: fake-devices-wrapper wrappre interface
Adds a new interface UbuntuDriversFakePCIDevicesInterface for
wrapping the `ubuntu-drivers list` command, as well as the
`ubuntu-drivers install` command, with the ubuntu-drivers-common
fake-devices-wrapper script. Pass `fake-pci-devices` on the kernel
command line to change to this interface.

The list command is moved to be run on the live environment because
the script depends on umockdev and gir1.2-umockdev-1.0 to be installed,
but it appears that the apt source tree is not yet configured enough to
install these packages. This is fine since ubuntu-drivers should behave
the same on the live systems as it does in the target system.

Modifications to make these packages available in the live and
target system are not in this commit.