drivers: debug flag to disable --gpgpu on server
On server, we force the usage of the --gpgpu flag in the ubuntu-drivers
invocation which has the effect of ignoring broadcom wireless drivers
at install time. To make MOK enrollment testing easier, this adds
a debug flag that disables this behavior. Pass `server-force-no-gpgpu`
on the kernel commandline to enable this.