filesystem: reset-partition size bug
In python, bool is a subclass of int so we can't use isinstance
to check if the user specified a size for the reset partition.
This causes autoinstall with "reset-partition: True" and
"reset-partition-only: true" to crash the installer due to creating
a reset-partition of size 1 (LP: #2061042).

(cherry picked from commit 652f73209c862860d219f60aaf742197cc98b56a)