storage: mark model config. when guided does not support customization
On core desktop, the user is not allowed to customize the storage layout
(i.e., perform further manual partitioning) after selecting the disk.

Therefore, when calling POST /storage/guided, the storage model
automatically gets marked configured.

However, when calling the v2 version of it (i.e., POST
/storage/v2/guided), the model did not get marked configured.

We now do but in the future we might want to reconsider this decision.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>