network: auto-disable dead interfaces on configure
It's possible that clients do not interact with the network controller -
even if the controller is interactive - and simply mark the controller
as configured (currently the desktop installer does this). Subiquity's
policy is to disable interfaces without a global IP address by the
time the network screen is shown. However, other than special handling
for autoinstall cases, this logic is only invoked the first time
the screen is shown. This makes sure the config written to the target
device adheres to policy.

This change purposefully does not apply the config in the live
environment. It's likely Desktop does not interact with the network
controller to avoid interactions between networkd and NetworkManager.
Handling of this logic should be for future commits.