storage: use probert to determine if firmware supports NVMe/TCP booting
Using the new probert-firmware, we now check if the firmware happens to
be EDK II built using the timberland-sig tree. Of course, this won't be
enough going forward because not all firmware that support NVMe/TCP
booting will be EDK II ; but it should be enough for the PoC with EDK2.

Alternatively, one can supply the --[no-]supports-nvme-tcp-booting to
act as if the firmware did/did not support NVMe/TCP booting.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>