storage: move NVMe/TCP stuff from the controller to the model
When determining if a given fsobj can be bootable, we need to know if
NVMe/TCP booting is supported. At the fsobj level, we don't have access
to the filesystem controller so calling supports_nvme_tcp_booting is not
an option. However, fsobj-s have access to the filesystem model.

We now move the `supports_nvme_tcp_booting` logic from the controller to
the model ; so that fsobj-s can access it.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>