storage: fix wrong debug log when opt_supports_nvme_tcp_booting is None
The following log line is supposed to be visible when the result of the
NVMe/TCP boot firmware detection does not match the
--[no]-supports-nvme-tcp-booting option.

  "but CLI argument states otherwise, so ignoring"

However, if no CLI option is passed, the log was visible as well, which
is not expected.

Fixed by printing the log line only if opt_supports_nvme_tcp_booting is
not None. None indicates the absence of the CLI option.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>