scripts: add script to inject ProbeData.txt into existing machine-config
The script can be used to inject a ProbeData.txt file into an existing
machine-config file.

By default, it will:
 * read the storage probe data from a file named ProbeData.txt in the
   current directory
 * read the machine-config from examples/machines/simple.json
 * write the resulting machine-config to stdout

The path to ProbeData.txt can be specified using the
--storage-probe-data option.
The patch to the machine-config can be specified using the
--machine-config option.
The --overwrite option can be passed to overwrite the original
machine-config.

examples:

 $ scripts/inject-probe-data.py \
    --storage-probe-data ProbeData.txt \
    --machine-config examples/machines/virtio.json
    --overwrite

 $ scripts/inject-probe-data.txt > simple-with-injected-storage.txt

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>