storage: add TODO to anticipate guided scenarios issues
Expect for manual partitioning, we should ideally suggest guided
scenarios only if we are strongly confident that they can be applied.

Currently, we only perform some basic checks before suggesting
scenarios. Going forward, I think we should be more cautious by trying
to apply the scenarios in a throwaway filesystem model. If the scenario
fails to apply, we should not suggest it.

There is a lot of work involved so this is marked as a TODO for now.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>