storage: don't suggest use-gap if we can't fit more partitions
If a disk has already reached its limit of (primary) partitions,
suggesting a use-gap scenario can only lead to failures. In the same
vein, if a disk only has room for one more partition but we also need to
add an ESP, the scenario would fail to apply too.

We now try to prevent Subiquity from suggesting such, bad, use-gap
scenarios ; by guessing the number of primary partitions needed ; and
comparing it with the limit.

Although it isn't a perfect solution (see below), this change should
hopefully address the most common problems.

What makes it difficult to properly suggest scenarios is that we do not
have all the cards in hand with the current API. For instance, the user
can choose to add a recovery partition, which of course influences the
number of partitions required. Sadly, we don't have a way with the
current API to allow a scenario without recovery partition ; while
rejecting the same scenario with a recovery partition. Capabilities can
also influence the number of required partitions, supposedly.

LP: #2073378

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 4f012e02008eb033a8d4d20e948c34588ac41c23)